gpucc/dts: sdmmagpie: Add 140MHz GPU frequency step
140MHz seems to be the lowest possible frequency to boot
and work with and is good enough for light loads such as
scrolling through applications without any lags.

Let's use a custom voltage level which is slightly higher
than the retention level and lower than the LOW SVS level,
and reduce the bus frequency as 140Mhz shouldn't be much
demanding, to reduce power consumption.

[ghostrider-reborn]
- Tested with a few days of normal usage and 3dmark runs at
140Mhz locked frequency, to see no crashes, artifacts or
GPU faults.

Co-authored-by: Forenche <prahul2003@gmail.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Forenche <prahul2003@gmail.com>
Change-Id: I027416cfc06cddf388f9cec14ad58d06ddac305d
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
[vbajs: Adapt with e9e83d3e3faa8b9a9df0f1ec639a2889409b28ce ("gpucc-sdmmagpie: Inline regulators") by simply revert-picking https://github.com/rd-stuffs/msm-4.14/commit/89b42ea28038987abf9cd0090235d8dd2aafa3f3 ]
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>