overlayfs:dont make DCACHE_OP_{HASH,COMPARE} weird
Recent Android deprecated sdcardfs and uses native fs casefold
functionality to achieve case insensitivity on /sdcard[1][2].

However, this breaks overlayfs on userdata partition because this
partition is now formatted with casefold option.

Overlayfs checks DCACHE_OP_{HASH,COMPARE} to ensure the filesystem is
case sensitive. However, this flags userdata partition as a false
positive. Most of the userdata partition is indeed case sensitive
except [3].

So this patch workarounds the issue by removing the check, which is
DANGEROUS! The responsibility to ensure fs case sensitivity is
shifted to the users.

Fixes overlayfs driver for docker.

[1]: https://source.android.com/docs/core/storage/sdcardfs-deprecate
[2]: https://android.googlesource.com/platform/system/vold/+/5b711b10dbbec26cd8157672f12566b525d0d2bb/model/PrivateVolume.cpp#170
[3]: https://android.googlesource.com/platform/system/vold/+/5b711b10dbbec26cd8157672f12566b525d0d2bb/model/PrivateVolume.cpp#178

Change-Id: Ia3a4a81e6772796b3400fd13b89c6f3677e35872
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>