arm64/lib: Handle CRC-32 alternative in C code
In preparation for adding another code path for performing CRC-32, move
the alternative patching for ARM64_HAS_CRC32 into C code. The logic for
deciding whether to use this new code path will be implemented in C too.

Adapted to 4.14: this tree's crc32.S never had the alternative_if_not
ARM64_HAS_CRC32 fallback blocks (the weak generic in lib/crc32.c is the
fallback), so this is a pure rename of the assembly symbols to *_arm64
plus a new crc32-glue.c providing crc32_le/__crc32c_le/crc32_be as C
wrappers. The <asm/alternative.h> include is dropped as it is now unused.

Change-Id: 28927cf8352f834277890a147fc3638db838a45b
Link: https://lore.kernel.org/r/20241018075347.2821102-6-ardb+git@google.com
Reviewed-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>