arm64/crc32: Implement 4-way interleave using PMULL
Now that kernel mode NEON no longer disables preemption, using FP/SIMD
in library code which is not obviously part of the crypto subsystem is
no longer problematic, as it will no longer incur unexpected latencies.

So accelerate the CRC-32 library code on arm64 to use a 4-way
interleave, using PMULL instructions to implement the folding.

On Apple M2, this results in a speedup of 2 - 2.8x when using input
sizes of 1k - 8k. For smaller sizes, the overhead of preserving and
restoring the FP/SIMD register file may not be worth it, so 1k is used
as a threshold for choosing this code path.

The coefficient tables were generated using code provided by Eric. [0]

[0] https://github.com/ebiggers/libdeflate/blob/master/scripts/gen_crc32_multipliers.c

Adapted to 4.14:
- cpu_have_named_feature(PMULL) is a 6.x API; use (elf_hwcap & HWCAP_PMULL)
  from <asm/hwcap.h> as the existing arch/arm64/crypto code does.
- crypto_simd_usable() does not exist here; use may_use_simd() from
  <asm/simd.h>. <crypto/internal/simd.h> is not needed and is omitted.

Change-Id: fbd391b916e7c4dd89d37f78440454b1da5523db
Link: https://lore.kernel.org/r/20241018075347.2821102-8-ardb+git@google.com
Cc: Eric Biggers <ebiggers@kernel.org>
Reviewed-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>