fuse: Replicate TASK_FREEZABLE behavior on downstream
msm-4.14 lacks TASK_FREEZABLE which was introduced with freezer rewrite in
upstream [1]. That's a huge backport that's not worth for the small change in
fuse. Instead, replicate how the logic worked before the introduction of
TASK_FREEZABLE in upstream and adapt for our kernel.

Now wait_event_state_exclusive() is only exclusive for freezable tasks. So,
let's edit the macro name to wait_event_freezable_state_exclusive().

[1] https://github.com/torvalds/linux/commit/f5d39b020809146

Change-Id: I5d4b277ee6cf3db2fb55228af18d44d4505f7a00
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>